dev: учет рекомендаций, причесывание кода, настройка сервера
