+++ app.js
@@ -6,39 +6,31 @@ const helmet = require('helmet');
 const mongoose = require('mongoose');
 const bodyParser = require('body-parser');
 const { errors, celebrate, Joi } = require('celebrate');
-const rateLimit = require('express-rate-limit');
+const limiter = require('./utils/config');
 const { errLogger, apiLogger } = require('./middlewares/logger');
+const errorHandler = require('./errors/errorHandler');
 
 const NotFoundError = require('./errors/notFoundError');
 
 const { createUser, loginUser } = require('./controllers/usersController');
 const auth = require('./middlewares/auth');
 
-/**
- * Защита от DDoS- атак
- */
-const limiter = rateLimit({
-  windowMs: 15 * 60 * 1000,
-  max: 1000,
-});
-
-const { PORT = 3000 } = process.env;
+const { PORT = 3000, DB_ADDRESS } = process.env;
 
 const app = express();
 
 app.use(helmet());
+app.use(cors());
 
 /**
  * Подключение к MongoDB
  */
-mongoose.connect('mongodb://localhost:27017/moviesexplorerdb', {
+mongoose.connect(DB_ADDRESS, {
   useNewUrlParser: true,
   useCreateIndex: true,
   useFindAndModify: false,
 })
   .then(() => console.log('Movies Explorer is connected to DB'));
-
-app.use(cors());
 
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
@@ -70,6 +62,7 @@ app.post('/signup',
 
 app.use('/', auth, require('./routes/users'));
 app.use('/', auth, require('./routes/movies'));
+
 app.use('*', (req, res, next) => {
   next(new NotFoundError('Страница не найдена'));
 });
@@ -77,19 +70,7 @@ app.use('*', (req, res, next) => {
 app.use(errLogger);
 app.use(errors());
 
-/**
- * Обработка ошибок
- */
-app.use((err, req, res, next) => {
-  const { message } = err;
-  const statusCode = err.statusCode || 500;
-  res.status(statusCode).send({
-    message: statusCode === 500
-      ? 'Произошла ошибка на сервере'
-      : message,
-  });
-  next();
-});
+app.use(errorHandler);
 
 app.listen(PORT, () => {
   console.log(`Movie Explorer Backend is listening on port ${PORT}`);
