+++ errors/errorHandler.js
@@ -1,0 +1,13 @@
+/**
+ * Обработчик ошибок
+ */
+module.exports = (err, req, res, next) => {
+  const { message } = err;
+  const statusCode = err.statusCode || 500;
+  res.status(statusCode).send({
+    message: statusCode === 500
+      ? 'Произошла ошибка на сервере'
+      : message,
+  });
+  next();
+};
