+++ package.json
@@ -25,12 +25,14 @@
     "cookie-parser": "^1.4.5",
     "cors": "^2.8.5",
     "dotenv": "^8.2.0",
+    "eslint-config-airbnb": "^18.2.1",
     "express": "^4.17.1",
     "express-rate-limit": "^5.2.6",
     "express-winston": "^4.1.0",
     "helmet": "^4.4.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.12.3",
+    "nodemon": "^2.0.7",
     "validator": "^13.5.2",
     "winston": "^3.3.3"
   }
