+++ routes/movies.js
@@ -19,12 +19,12 @@ router.post('/movies',
       country: Joi.string().required(),
       director: Joi.string().required(),
       duration: Joi.number().required(),
-      year: Joi.string().required(),
+      year: Joi.string().required().min(4).length(4),
       description: Joi.string().required(),
       image: Joi.string().required().pattern(regex),
       trailer: Joi.string().required().pattern(regex),
       thumbnail: Joi.string().required().pattern(regex),
-      movieId: Joi.number().required(),
+      movieId: Joi.number().positive().required().integer(),
       nameRU: Joi.string().required(),
       nameEN: Joi.string().required(),
     }),
