+++ utils/config.js
@@ -1,0 +1,11 @@
+/**
+ * Защита от DDoS- атак
+ */
+const rateLimit = require('express-rate-limit');
+
+const limiter = rateLimit({
+  windowMs: 15 * 60 * 1000,
+  max: 1000,
+});
+
+module.exports = limiter;
